requirements.txt
Flask
Flask-Cors
python-dotenv
requests
scikit-learn
sentence-transformers
torch
transformers
pylint
isort
